pyproject.toml
[tool.poetry]
name = "tempus_flask"
version = "0.1.0"
description = ""
authors = ["Ed Jones <edjones1999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
flask-restful = "^0.3.8"
flask-classy = "^0.6.10"
flask-migrate = "^2.5.3"
python-dotenv = "^0.12.0"
flask-wtf = "^0.14.3"
flask-sqlacodegen = "^1.1.7"
psycopg2 = "^2.8.4"
flask-marshmallow = "^0.11.0"
marshmallow-sqlalchemy = "^0.22.3"
alembic = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
